+++ src/routes/ManageAds/hooks/index.ts
@@ -1,2 +1,1 @@
 export { useFetchAdsQuery } from './useFetchAdsQuery'
-export { useFilterAds } from './useFilterAds'
